huge refactoring and old code deletion
