+++ .gitignore
@@ -1,4 +1,5 @@
 diploma_thesis/data
+diploma_thesis/not_working_pubtator_annotate_raw
 
 neo4j/data
 neo4j/import
