--- diploma_thesis/data/.gitattributes
@@ -1,1 +1,0 @@
-neo4j.dump filter=lfs diff=lfs merge=lfs -text
